+++ src/ESPRotary.cpp
@@ -47,6 +47,7 @@ void ESPRotary::begin(byte pin1, byte pin2, byte steps_per_click /* = 1 */, int 
 
   loop();
   steps = initial_pos * steps_per_click;
+  last_steps = steps;
   last_event = rotary_event::none;
   dir = rotary_direction::undefined;
 }
@@ -125,6 +126,7 @@ void ESPRotary::resetPosition(int p /* = 0 */, bool fireCallback /* = true */) {
   // yes...
   steps = p * steps_per_click;
   _isWithinBounds();
+  last_steps = steps;
   if (fireCallback) _callCallback(change_cb);
   last_event = rotary_event::none;
   dir = rotary_direction::undefined;
@@ -199,6 +201,12 @@ void* ESPRotary::getContext() const {
 
 /////////////////////////////////////////////////////////////////
 
+void ESPRotary::setPinReadFunction(PinReadFunction f) {
+  pinReadFn = (f != NULL) ? f : digitalRead;
+}
+
+/////////////////////////////////////////////////////////////////
+
 bool ESPRotary::operator==(const ESPRotary& rhs) const {
   return (this == &rhs);
 }
@@ -225,7 +233,7 @@ void ESPRotary::loop() {
 
 bool ESPRotary::_wasRotated() {
   static const int8_t factors[] = {0, 1, -1, 2, -1, 0, -2, 1, 1, -2, 0, -1, 2, -1, 1, 0};
-  int encoderState = (state & 3) | digitalRead(pin1) << 2 | digitalRead(pin2) << 3 ;
+  int encoderState = (state & 3) | pinReadFn(pin1) << 2 | pinReadFn(pin2) << 3 ;
   steps += factors[encoderState] * increment;
   state = (encoderState >> 2);
   int stepDifference = abs(steps - last_steps);